% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasttext_interface.R
\name{compute_elapsed_time}
\alias{compute_elapsed_time}
\title{elapsed time in hours & minutes & seconds}
\usage{
compute_elapsed_time(time_start)
}
\arguments{
\item{time_start}{a numeric value specifying the start time}
}
\value{
It does not return a value but only prints the time in form of a character string in the R session
}
\description{
elapsed time in hours & minutes & seconds
}
\keyword{internal}
