% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{give_args_fasttext}
\alias{give_args_fasttext}
\title{The Rcpp function which is used in the 'fasttext_interface' R function}
\usage{
give_args_fasttext(args, pth = "", MilliSecs = 100L, pth_in = "",
  queryWord = "", remove_previous_file = TRUE)
}
\description{
The Rcpp function which is used in the 'fasttext_interface' R function
}
\keyword{internal}
