% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasttext_interface.R
\name{inner_elapsed_time}
\alias{inner_elapsed_time}
\title{inner function of 'compute_elapsed_time'}
\usage{
inner_elapsed_time(secs, estimated = FALSE)
}
\arguments{
\item{secs}{a numeric value specifying the seconds}

\item{estimated}{a boolean. If TRUE then the output label becomes the 'Estimated time'}
}
\value{
a character string showing the estimated or elapsed time
}
\description{
inner function of 'compute_elapsed_time'
}
\keyword{internal}
