% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{printAnalogiesUsage}
\alias{printAnalogiesUsage}
\title{Print Usage Information when the command equals to 'analogies'}
\usage{
printAnalogiesUsage()
}
\description{
Print Usage Information when the command equals to 'analogies'
}
\examples{

library(fastText)

printAnalogiesUsage()

}
