% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{printDumpUsage}
\alias{printDumpUsage}
\title{Print Usage Information when the command equals to 'dump'}
\usage{
printDumpUsage()
}
\description{
Print Usage Information when the command equals to 'dump'
}
\examples{

library(fastText)

printDumpUsage()

}
