% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{printNNUsage}
\alias{printNNUsage}
\title{Print Usage Information when the command equals to 'nn'}
\usage{
printNNUsage()
}
\description{
Print Usage Information when the command equals to 'nn'
}
\examples{

library(fastText)

printNNUsage()

}
