% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{printPrintNgramsUsage}
\alias{printPrintNgramsUsage}
\title{Print Usage Information when the command equals to 'print-ngrams'}
\usage{
printPrintNgramsUsage()
}
\description{
Print Usage Information when the command equals to 'print-ngrams'
}
\examples{

library(fastText)

printPrintNgramsUsage()

}
