% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{printPrintSentenceVectorsUsage}
\alias{printPrintSentenceVectorsUsage}
\title{Print Usage Information when the command equals to 'print-sentence-vectors'}
\usage{
printPrintSentenceVectorsUsage()
}
\description{
Print Usage Information when the command equals to 'print-sentence-vectors'
}
\examples{

library(fastText)

printPrintSentenceVectorsUsage()

}
