% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{printPrintWordVectorsUsage}
\alias{printPrintWordVectorsUsage}
\title{Print Usage Information when the command equals to 'print-word-vectors'}
\usage{
printPrintWordVectorsUsage()
}
\description{
Print Usage Information when the command equals to 'print-word-vectors'
}
\examples{

library(fastText)

printPrintWordVectorsUsage()

}
