% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{printQuantizeUsage}
\alias{printQuantizeUsage}
\title{Print Usage Information when the command equals to 'quantize'}
\usage{
printQuantizeUsage()
}
\description{
Print Usage Information when the command equals to 'quantize'
}
\examples{

library(fastText)

printQuantizeUsage()

}
