% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{printTestLabelUsage}
\alias{printTestLabelUsage}
\title{Print Usage Information when the command equals to 'test-label'}
\usage{
printTestLabelUsage()
}
\description{
Print Usage Information when the command equals to 'test-label'
}
\examples{

library(fastText)

printTestLabelUsage()

}
