% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{printTestUsage}
\alias{printTestUsage}
\title{Print Usage Information when the command equals to 'test'}
\usage{
printTestUsage()
}
\description{
Print Usage Information when the command equals to 'test'
}
\examples{

library(fastText)

printTestUsage()

}
