% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{printUsage}
\alias{printUsage}
\title{Print Usage Information for all parameters}
\usage{
printUsage()
}
\description{
Print Usage Information for all parameters
}
\examples{

library(fastText)

printUsage()

}
