% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasttext_interface.R
\name{multiplot}
\alias{multiplot}
\title{Multiple plot function}
\usage{
multiplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{ellipsis to pass ggplot objects}

\item{plotlist}{either NULL or a list of ggplot objects}

\item{cols}{Number of columns in layout}

\item{layout}{A matrix specifying the layout. If present, 'cols' is ignored}
}
\value{
It does not return a value but only shows the ggplots in the R session
}
\description{
Multiple plot function
}
\details{
If the layout is something like matrix(c(1,2,3,3), nrow = 2, byrow = TRUE), then plot 1 will
go in the upper left, 2 will go in the upper right, and 3 will go all the way across the bottom.
}
\references{
http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/
}
\keyword{internal}
