% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasttext_interface.R
\name{print_parameters}
\alias{print_parameters}
\title{Print the parameters for a specific command}
\usage{
print_parameters(command = "supervised")
}
\arguments{
\item{command}{a character string specifying the command for which the parameters should be printed in the R session}
}
\value{
It does not return a value but only prints the available parameters in the R session
}
\description{
Print the parameters for a specific command
}
\examples{

\dontrun{

library(fastText)

print_parameters(command = 'supervised')
}
}
\references{
https://github.com/facebookresearch/fastText#full-documentation

https://github.com/facebookresearch/fastText/issues/341#issuecomment-339783130
}
