% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{printAnalogiesUsage}
\alias{printAnalogiesUsage}
\title{Print Usage Information when the command equals to 'analogies'}
\usage{
printAnalogiesUsage(verbose = TRUE)
}
\arguments{
\item{verbose}{if TRUE then information will be printed in the console}
}
\value{
It does not return a value but only prints the available parameters of the 'printAnalogiesUsage' function in the R session
}
\description{
Print Usage Information when the command equals to 'analogies'
}
\examples{

library(fastText)

printAnalogiesUsage()

}
