% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{printDumpUsage}
\alias{printDumpUsage}
\title{Print Usage Information when the command equals to 'dump'}
\usage{
printDumpUsage(verbose = TRUE)
}
\arguments{
\item{verbose}{if TRUE then information will be printed in the console}
}
\value{
It does not return a value but only prints the available parameters of the 'printDumpUsage' function in the R session
}
\description{
Print Usage Information when the command equals to 'dump'
}
\examples{

library(fastText)

printDumpUsage()

}
