% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{printPredictUsage}
\alias{printPredictUsage}
\title{Print Usage Information when the command equals to 'predict' or 'predict-prob'}
\usage{
printPredictUsage()
}
\value{
It does not return a value but only prints the available parameters of the 'printPredictUsage' function in the R session
}
\description{
Print Usage Information when the command equals to 'predict' or 'predict-prob'
}
\examples{

library(fastText)

printPredictUsage()

}
