% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{printPrintWordVectorsUsage}
\alias{printPrintWordVectorsUsage}
\title{Print Usage Information when the command equals to 'print-word-vectors'}
\usage{
printPrintWordVectorsUsage()
}
\value{
It does not return a value but only prints the available parameters of the 'printPrintWordVectorsUsage' function in the R session
}
\description{
Print Usage Information when the command equals to 'print-word-vectors'
}
\examples{

library(fastText)

printPrintWordVectorsUsage()

}
