% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{printTestLabelUsage}
\alias{printTestLabelUsage}
\title{Print Usage Information when the command equals to 'test-label'}
\usage{
printTestLabelUsage(verbose = TRUE)
}
\arguments{
\item{verbose}{if TRUE then information will be printed in the console}
}
\value{
It does not return a value but only prints the available parameters of the 'printTestLabelUsage' function in the R session
}
\description{
Print Usage Information when the command equals to 'test-label'
}
\examples{

library(fastText)

printTestLabelUsage()

}
