% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{printTestUsage}
\alias{printTestUsage}
\title{Print Usage Information when the command equals to 'test'}
\usage{
printTestUsage()
}
\value{
It does not return a value but only prints the available parameters of the 'printTestUsage' function in the R session
}
\description{
Print Usage Information when the command equals to 'test'
}
\examples{

library(fastText)

printTestUsage()

}
