% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{printUsage}
\alias{printUsage}
\title{Print Usage Information for all parameters}
\usage{
printUsage(verbose = TRUE)
}
\arguments{
\item{verbose}{if TRUE then information will be printed in the console}
}
\value{
It does not return a value but only prints the available parameters of the 'printUsage' function in the R session
}
\description{
Print Usage Information for all parameters
}
\examples{

library(fastText)

printUsage()

}
